@@ -14,7 +14,7 @@
    limitations under the License.
  */
 import { Injectable } from '@angular/core';
-import { Http, Response, RequestOptionsArgs, Headers } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { PlatformLocation } from '@angular/common';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
@@ -30,7 +30,7 @@ import { OrderbookBs } from '../common/orderbookBs';
 @Injectable()
 export class BitstampService {    
    
-    private _reqOptionsArgs: RequestOptionsArgs = { headers: new Headers() };
+    private _reqOptionsArgs = { headers: new HttpHeaders().set( 'Content-Type', 'application/json' ) };
     private readonly _bitstamp = '/bitstamp';    
     BTCEUR = 'btceur';
     ETHEUR = 'etheur';
@@ -42,19 +42,18 @@ export class BitstampService {
     XRPUSD = 'xrpusd';
     private _utils = new Utils(); 
 
-    constructor(private http: Http, private pl: PlatformLocation ) { 
-        this._reqOptionsArgs.headers.set( 'Content-Type', 'application/json' );
+    constructor(private http: HttpClient, private pl: PlatformLocation ) { 
     }
 
     getCurrentQuote(currencypair: string): Observable<QuoteBs> {
-        return this.http.get(this._bitstamp+'/'+currencypair+'/current', this._reqOptionsArgs).map(res => <QuoteBs>res.json()).catch(this._utils.handleError);
+        return this.http.get(this._bitstamp+'/'+currencypair+'/current', this._reqOptionsArgs).catch(this._utils.handleError);
     }
      
     getTodayQuotes(currencypair: string): Observable<QuoteBs[]> {
-        return this.http.get(this._bitstamp+'/'+currencypair+'/today', this._reqOptionsArgs).map(res => <QuoteBs[]>res.json()).catch(this._utils.handleError);
+        return this.http.get(this._bitstamp+'/'+currencypair+'/today', this._reqOptionsArgs).catch(this._utils.handleError);
     }
     
     getOrderbook(currencypair: string): Observable<OrderbookBs> {
-        return this.http.get(this._bitstamp+'/'+currencypair+'/orderbook/', this._reqOptionsArgs).map(res => <OrderbookBs>res.json()).catch(this._utils.handleError);
+        return this.http.get(this._bitstamp+'/'+currencypair+'/orderbook/', this._reqOptionsArgs).catch(this._utils.handleError);
     }
 }
